Combine Cellss
Attempting to add in combination of cells. Do this by checking if the sum of the neighbors is larger than zero. If it is then add the largest neighbor and set position equal to 0,0 to remove itself from simulation. Done by checking if positon = 0,0 then continue through loop

Problem arises where I think it it is still moving the ones that are removed, also should maybe keep track of size