First Attempt at infection
So first attempt at this and it is bugged. Infection within groups seem to be disappear and so are clusters. I wonder if two clusters are trying to combine into one another and both set themselves to zero. Maybe should adjust it so that instead of moving cluster into it's neighbor, have the neighbor move into it?

Changed the clusters to update from the cluster size list instead of the environment in the beginning and throughout. Many clusters started disappearing from cluster size list and infection wasn't updating correctly. lines 137-146 is trying to update infection and you can see display statements for debug. what's strange is some of them don't even output anything, when they should be outputting zeros so that may be worth looking into.

lines 33-35 help setup how infection should be done

line 47 sets up cluster size to hold the size in position 1 and the number of infected amoeba in position 2.

63-68 Initialize the size list of infected, currently all amoebas in a cluster are infected at once.

lines 150 and 173 both update the size and infection status of the current cluster for the position list